Moved serialization function to model
