+++ src/controllers/user/data/types.ts
@@ -1,10 +1,3 @@
 export type GenericObject<m = any> = {
   [key: string]: m;
 };
-
-export type APIUser = {
-  name: string;
-  email: string;
-  avatar: string;
-  id: string;
-};
